refactor: 이너 예외 클래스 분리
