+++ src/main/java/roomescape/application/auth/AuthService.java
@@ -5,7 +5,7 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.application.auth.dto.JwtPayload;
 import roomescape.application.auth.dto.LoginParam;
 import roomescape.application.auth.dto.LoginResult;
-import roomescape.application.support.exception.UnauthorizedException.LoginAuthException;
+import roomescape.application.support.exception.LoginAuthException;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
