+++ src/main/java/roomescape/application/support/exception/ForbiddenException.java
@@ -1,6 +1,7 @@
 package roomescape.application.support.exception;
 
 public class ForbiddenException extends CoreException {
+
     public ForbiddenException(String message) {
         super(message);
     }
