+++ src/main/java/roomescape/application/support/exception/LoginAuthException.java
@@ -1,0 +1,12 @@
+package roomescape.application.support.exception;
+
+public class LoginAuthException extends UnauthorizedException {
+
+    public LoginAuthException(String message) {
+        super(message);
+    }
+
+    public LoginAuthException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
