+++ src/main/java/roomescape/application/support/exception/UnauthorizedException.java
@@ -9,15 +9,4 @@ public class UnauthorizedException extends CoreException {
     public UnauthorizedException(String message, Throwable cause) {
         super(message, cause);
     }
-
-    public static class LoginAuthException extends UnauthorizedException {
-
-        public LoginAuthException(String message) {
-            super(message);
-        }
-
-        public LoginAuthException(String message, Throwable cause) {
-            super(message, cause);
-        }
-    }
 }
