+++ src/main/java/roomescape/presentation/support/AuthInfoResolveException.java
@@ -1,0 +1,12 @@
+package roomescape.presentation.support;
+
+public class AuthInfoResolveException extends ParameterResolveException {
+
+    public AuthInfoResolveException(String message) {
+        super(message);
+    }
+
+    public AuthInfoResolveException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
