+++ src/test/java/roomescape/application/auth/AuthServiceTest.java
@@ -11,7 +11,7 @@ import roomescape.application.AbstractServiceIntegrationTest;
 import roomescape.application.auth.dto.JwtPayload;
 import roomescape.application.auth.dto.LoginParam;
 import roomescape.application.auth.dto.LoginResult;
-import roomescape.application.support.exception.UnauthorizedException.LoginAuthException;
+import roomescape.application.support.exception.LoginAuthException;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
